contrib/net/http: move htto client short-hands into a dedicated client sub-backage
