+++ contrib/net/http/orchestrion.client.yml
@@ -192,18 +192,18 @@ aspects:
       # Wire the context that is found to the handlers...
       - wrap-expression:
           imports:
-            httptrace: gopkg.in/DataDog/dd-trace-go.v1/contrib/net/http
+            client: gopkg.in/DataDog/dd-trace-go.v1/contrib/net/http/client
           template: |-
             {{- $ctx := .Function.ArgumentOfType "context.Context" -}}
             {{- $req := .Function.ArgumentOfType "*net/http.Request" }}
             {{- if $ctx -}}
-              httptrace.{{ .AST.Fun.Name }}(
+              client.{{ .AST.Fun.Name }}(
                 {{ $ctx }},
                 {{ range .AST.Args }}{{ . }},
                 {{ end }}
               )
             {{- else if $req -}}
-              httptrace.{{ .AST.Fun.Name }}(
+              client.{{ .AST.Fun.Name }}(
                 {{ $req }}.Context(),
                 {{ range .AST.Args }}{{ . }},
                 {{ end }}
